Add files via upload
Recommendations on which characteristics are most important for determining whether a customer has a favourable perception of the microvan concept, to segment the market, and to determine which segment(s) would be good to target based on the analysis of data.

Gross Pointe Associates (GPA) seeks a deeper understanding of various automobile attributes and psychographic variables that might influence consumer perception of the microvan. GPA has collected 400 responses from the survey outlined in the case. The data is posted as “microvan.dta”